[BACKLOG-13281] - Mavenized xml-input-stream-plugin
