+++ plugins/xml-input-stream-plugin/.gitignore
@@ -1,0 +1,4 @@
+bin
+.idea
+target
+*.iml
