+++ plugins/xml-input-stream-plugin/assemblies/pom.xml
@@ -1,0 +1,19 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<project xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd" xmlns="http://maven.apache.org/POM/4.0.0"
+         xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance">
+    <modelVersion>4.0.0</modelVersion>
+    <parent>
+        <groupId>org.pentaho.kettle.plugins</groupId>
+        <artifactId>xml-input-stream-plugin-parent</artifactId>
+        <version>8.0-SNAPSHOT</version>
+    </parent>
+    <groupId>org.pentaho.kettle.plugins</groupId>
+    <artifactId>xml-input-stream-plugin-assemblies</artifactId>
+    <version>8.0-SNAPSHOT</version>
+    <packaging>pom</packaging>
+    <name>${project.artifactId}</name>
+    <description></description>
+    <modules>
+        <module>plugin</module>
+    </modules>
+</project>
