+++ plugins/xml-input-stream-plugin/build-res/subfloor-pkg.xml
