+++ plugins/xml-input-stream-plugin/core/src/main/java/org/pentaho/di/trans/steps/xmlinputsax/messages/messages_ja_JP.properties
