+++ plugins/xml-input-stream-plugin/core/src/main/resources/org/pentaho/di/trans/steps/xmlinputsax/messages/messages_en_US.properties
@@ -1,0 +1,67 @@
+XMLInputSaxDialog.Button.PreviewRows.Label=\ \ &Preview rows   
+XMLInputSaxDialog.Shell.Text=Streaming XML Input
+XMLInputSaxDialog.StepName.Label=Step name 
+XMLInputSaxDialog.File.Label=File
+XMLInputSaxDialog.FileOrDirectory.Label=File or directory 
+XMLInputSaxDialog.System.Button.Browse=&Browse
+XMLInputSaxDialog.Button.Brows.ToolTip=Browse for a file or directory & add to the list
+XMLInputSaxDialog.Button.Variable.Label=&Variable
+XMLInputSaxDialog.Button.Variable.ToolTip=Insert a variable in the filename or directory
+XMLInputSaxDialog.Button.AddFile.Label=&Add
+XMLInputSaxDialog.Button.AddFile.ToolTip=Add this entry to the list of files & directories.
+XMLInputSaxDialog.FileMaskRegExp.Label=Regular Expression 
+XMLInputSaxDialog.SelectedFiles.Label=Selected files: 
+XMLInputSaxDialog.Button.DeleteEntry.Label=&Delete
+XMLInputSaxDialog.Button.DeleteEntry.ToolTip=Delete the selected entries from the list of files.
+XMLInputSaxDialog.Button.EditEntry.Label=&Edit
+XMLInputSaxDialog.Button.EditEntry.ToolTip=Edit the selected file and remove from the list.
+XMLInputSaxDialog.Button.ShowFilenames.Label=&Show filename(s)...
+XMLInputSaxDialog.Column.FileDirectory.Label=File/Directory
+XMLInputSaxDialog.Column.Wildcard.Label=Wildcard (RegExp)
+XMLInputSaxDialog.Column.FileDirectory.ToolTip=Enter a regular expression here and a directory in the first column.
+XMLInputSaxDialog.Tab.Content.Label=Content
+XMLInputSaxDialog.IncludeFilename.Label=Include filename in output? 
+XMLInputSaxDialog.IncludeFilename.ToolTip=Check this to add a field (String) containing the filename.
+XMLInputSaxDialog.FilenameFieldname.Label=Filename fieldname 
+XMLInputSaxDialog.RowNumInOutput.Label=Rownum in output? 
+XMLInputSaxDialog.RowNumInOutput.ToolTip=Check this to add a field (String) containing the filename.
+XMLInputSaxDialog.RowNumInOutputField.Label=Rownum fieldname 
+XMLInputSaxDialog.Limit.Label=Limit 
+XMLInputSaxDialog.Location.ToolTip=Element/position\n------------------------\nA: still specify an attribute\nEp: specify an element defined by its position (equivalent to E in original XMLInput).\nEa: specify an element defined by an attribute and allow value parsing.\n\nExample:\n-------------------\nEp=element/1 this is the first element called ''element'\nEa=element/att:val this is the element called ''element'' that have an attribute called ''att'' with ''val'' value\n
+XMLInputSaxDialog.Location.Label=Location (see also: tooltip help)
+XMLInputSaxDialog.Column.Elements.Label=Elements
+XMLInputSaxDialog.Tab.Fields.Label=Fields
+XMLInputSaxDialog.Button.GetFields.Label=\ &Get fields 
+XMLInputSaxDialog.Column.Name.Label=Name
+XMLInputSaxDialog.Column.Type.Label=Type
+XMLInputSaxDialog.Column.Format.Label=Format
+XMLInputSaxDialog.Column.Length.Label=Length
+XMLInputSaxDialog.Column.Precision.Label=Precision
+XMLInputSaxDialog.Column.Currency.Label=Currency
+XMLInputSaxDialog.Column.Decimal.Label=Decimal
+XMLInputSaxDialog.Column.Group.Label=Group
+XMLInputSaxDialog.Column.TrimType.Label=Trim type
+XMLInputSaxDialog.Column.Repeat.Label=Repeat
+XMLInputSaxDialog.Column.Position.Label=Position
+XMLInputSaxDialog.Column.Element.Label=Element
+XMLInputSaxDialog.Column.DefiningAttribute.Label=Defining Attribute
+XMLInputSaxDialog.System.Button.OK=\ &OK 
+XMLInputSaxDialog.System.Button.Cancel=\ &Cancel 
+XMLInputSaxDialog.Dialog.FilesRead.Title=Files read
+XMLInputSaxDialog.Dialog.FilesRead.Message=Files read:
+XMLInputSaxDialog.Dialog.NoFilesFound.Message=No files found\!  Please check the filename/directory and regular expression options.
+XMLInputSaxDialog.Dialog.NoFilesFound.Title=ERROR
+XMLInputSaxDialog.Dialog.ErrorParsingInputData.Title=Error parsing input data
+XMLInputSaxDialog.Dialog.ErrorParsingInputData.Message=An error occurred while parsing the input data on this dialog
+XMLInputSaxDialog.Dialog.SelectVariable.Title=Select an Environment Variable
+XMLInputSaxDialog.Dialog.SelectVariable.Message=Select an Environment Variable
+XMLInputSaxDialog.60=XML files
+XMLInputSaxDialog.68=All files
+XMLInputSaxDialog.Dialog.EnterPreviewSize.Title=Enter preview size
+XMLInputSaxDialog.Dialog.EnterPreviewSize.Message=Enter the number of rows you would like to preview:
+XMLInputSaxDialog.Dialog.ErrorDisplayingPreviewData.Title=Error displaying preview data
+XMLInputSaxDialog.Dialog.ErrorDisplayingPreviewData.Message=An error occurred while trying to preview XML data
+XMLInputSaxDialog.Dialog.NoElements.Message=No element(s) defined.
+XMLInputSaxDialog.Dialog.NoElements.Title=Unable to get fields
+XMLInputSaxDialog.Name=Streaming XML Input
+XMLInputSaxDialog.Description=Read data from an XML file in a streaming fashing, working faster and consuming less memory
