+++ plugins/xml-input-stream-plugin/core/src/main/resources/org/pentaho/di/trans/steps/xmlinputsax/messages/messages_es_AR.properties
@@ -1,0 +1,65 @@
+XMLInputSaxDialog.Button.PreviewRows.Label=\ \ &Previsualizar filas   
+XMLInputSaxDialog.Shell.Text=XML contínuo - Entrada
+XMLInputSaxDialog.StepName.Label=Nombre paso 
+XMLInputSaxDialog.File.Label=Fichero
+XMLInputSaxDialog.FileOrDirectory.Label=Fichero o directorio 
+XMLInputSaxDialog.System.Button.Browse=E&xaminar
+XMLInputSaxDialog.Button.Brows.ToolTip=Buscar un fichero o directorio y añadir a la lista
+XMLInputSaxDialog.Button.Variable.Label=&Variable
+XMLInputSaxDialog.Button.Variable.ToolTip=Insertar una variable en el nombre de fichero o directorio
+XMLInputSaxDialog.Button.AddFile.Label=&Añadir
+XMLInputSaxDialog.Button.AddFile.ToolTip=Añadir esta entrada a la lista de ficheros y directorios.
+XMLInputSaxDialog.FileMaskRegExp.Label=Expresión Regular 
+XMLInputSaxDialog.SelectedFiles.Label=Ficheros seleccionados: 
+XMLInputSaxDialog.Button.DeleteEntry.Label=E&liminar
+XMLInputSaxDialog.Button.DeleteEntry.ToolTip=Eliminar las entradas seleccionadas de la lista de ficheros.
+XMLInputSaxDialog.Button.EditEntry.Label=&Editar
+XMLInputSaxDialog.Button.EditEntry.ToolTip=Editar el fichero seleccionado y eliminar de la lista.
+XMLInputSaxDialog.Button.ShowFilenames.Label=Mo&strar nombres de ficheros...
+XMLInputSaxDialog.Column.FileDirectory.Label=Fichero/Directorio
+XMLInputSaxDialog.Column.Wildcard.Label=Comodín
+XMLInputSaxDialog.Column.FileDirectory.ToolTip=Introduce una expresión regular aquí y un directorio en la primera columna.
+XMLInputSaxDialog.Tab.Content.Label=Contenido
+XMLInputSaxDialog.IncludeFilename.Label=¿Incluir nombre de fichero en salida? 
+XMLInputSaxDialog.IncludeFilename.ToolTip=Marca esto para añadir un campo (String) con el nombre de fichero.
+XMLInputSaxDialog.FilenameFieldname.Label=Nombre fichero nombre campo 
+XMLInputSaxDialog.RowNumInOutput.Label=¿Número de fila en salida? 
+XMLInputSaxDialog.RowNumInOutput.ToolTip=Marca esto para añadir un campo (String) con el número de fila.
+XMLInputSaxDialog.RowNumInOutputField.Label=Número fila nombre campo 
+XMLInputSaxDialog.Limit.Label=Límite 
+XMLInputSaxDialog.Location.ToolTip=Elemento/posición\n------------------------\nA: especifica un atributo\nEp: especifica un elemento definido por su posición (equivalente a E en XMLInput original).\nEa: especifica un elemento definido por un atributo y permitir analizado.\n\nEjemplo:\n-------------------\nEp=elemento/1 este es el primer elemento llamado ''element'\nEa=elemento/atr:val este es el elemento llamado ''element'' que tiene un atributo llamado ''atr'' con valor ''val'\n
+XMLInputSaxDialog.Location.Label=Ubicación (ver ayuda)
+XMLInputSaxDialog.Column.Elements.Label=Elementos
+XMLInputSaxDialog.Tab.Fields.Label=Campos
+XMLInputSaxDialog.Button.GetFields.Label=\ &Obtener campos 
+XMLInputSaxDialog.Column.Name.Label=Nombre
+XMLInputSaxDialog.Column.Type.Label=Tipo
+XMLInputSaxDialog.Column.Format.Label=Formato
+XMLInputSaxDialog.Column.Length.Label=Longitud
+XMLInputSaxDialog.Column.Precision.Label=Precisión
+XMLInputSaxDialog.Column.Currency.Label=Moneda
+XMLInputSaxDialog.Column.Decimal.Label=Decimal
+XMLInputSaxDialog.Column.Group.Label=Grupo
+XMLInputSaxDialog.Column.TrimType.Label=Tipo de poda
+XMLInputSaxDialog.Column.Repeat.Label=Repetir
+XMLInputSaxDialog.Column.Position.Label=Posición
+XMLInputSaxDialog.Column.Element.Label=Elemento
+XMLInputSaxDialog.Column.DefiningAttribute.Label=Definiendo Atributo
+XMLInputSaxDialog.System.Button.OK=\ &Vale 
+XMLInputSaxDialog.System.Button.Cancel=\ &Cancelar 
+XMLInputSaxDialog.Dialog.FilesRead.Title=Ficheros leídos
+XMLInputSaxDialog.Dialog.FilesRead.Message=Ficheros leídos:
+XMLInputSaxDialog.Dialog.NoFilesFound.Message=No se han encontrado ficheros\!  Por favor revisa el nombre del fichero/directorio y la expresión regular.
+XMLInputSaxDialog.Dialog.NoFilesFound.Title=ERROR
+XMLInputSaxDialog.Dialog.ErrorParsingInputData.Title=Error leyendo datos
+XMLInputSaxDialog.Dialog.ErrorParsingInputData.Message=Se ha producido un error leyendo los datos de entrada
+XMLInputSaxDialog.Dialog.SelectVariable.Title=Selecciona una Variable de Entorno
+XMLInputSaxDialog.Dialog.SelectVariable.Message=Selecciona una Variable de Entorno
+XMLInputSaxDialog.60=Ficheros XML
+XMLInputSaxDialog.68=Todos los ficheros
+XMLInputSaxDialog.Dialog.EnterPreviewSize.Title=Tamaño de previsualización
+XMLInputSaxDialog.Dialog.EnterPreviewSize.Message=Introduce el número de filas que quieres previsualizar
+XMLInputSaxDialog.Dialog.ErrorDisplayingPreviewData.Title=Error mostrando previsualización
+XMLInputSaxDialog.Dialog.ErrorDisplayingPreviewData.Message=Se ha producido un error al intentar previsualizar datos XML
+XMLInputSaxDialog.Name=Entrada XML Continuo
+XMLInputSaxDialog.Description=Leer datos de un fichero XML de modo cont\u00EDnuo, m\u00E1s r\u00E1pido y consumiendo menos memoria
