+++ plugins/xml-input-stream-plugin/core/src/main/resources/org/pentaho/di/trans/steps/xmlinputsax/messages/messages_fr_FR.properties
@@ -1,0 +1,71 @@
+#File generated by Pentaho Translator for package 'org.pentaho.di.trans.steps.xmlinputsax' in locale 'fr_FR'
+#
+#
+#Mon Nov 12 18:32:34 CET 2007
+XMLInputSaxDialog.Dialog.NoElements.Title=Impossible de r\u00E9cup\u00E9rer les champs
+XMLInputSaxDialog.Column.Precision.Label=Pr\u00E9cision
+XMLInputSaxDialog.Button.Variable.Label=&Variable
+XMLInputSaxDialog.Button.AddFile.Label=&Ajouter
+XMLInputSaxDialog.Column.Format.Label=Format
+XMLInputSaxDialog.Button.GetFields.Label=\ &R\u00E9cup\u00E9rer champs 
+XMLInputSaxDialog.Column.Name.Label=Nom
+XMLInputSaxDialog.System.Button.Cancel=\ &Annuler 
+XMLInputSaxDialog.Column.Elements.Label=\u00C9l\u00E9ments
+XMLInputSaxDialog.Column.Decimal.Label=D\u00E9cimal
+XMLInputSaxDialog.StepName.Label=Nom \u00E9tape 
+XMLInputSaxDialog.Dialog.SelectVariable.Title=S\u00E9lectionner une variable d''environnement
+XMLInputSaxDialog.Column.TrimType.Label=Trim type
+XMLInputSaxDialog.Button.Variable.ToolTip=Ins\u00E9rer une variable d''environnement dans le nom du fichier ou du r\u00E9pertoire
+XMLInputSaxDialog.Location.ToolTip=\u00C9l\u00E9ment/position\n------------------------\nA\: indiquez toujours un attribut\nEp\: indiquez un \u00E9l\u00E9ment d\u00E9fini par sa position (\u00E9quivalent \u00E0 E dans le flux original XML).\nEa\: indiquez un \u00E9l\u00E9ment d\u00E9fini par un attribut et autoris\u00E9 l''analyse.\n\nExemple\:\n-------------------\nEp\=element/1 c''est le premier \u00E9l\u00E9ment nomm\u00E9 ''element'\nEa\=element/att\:val c''est l'\u00E9l\u00E9ment nomm\u00E9 ''element'' qui a un attribut nomme ''att'' avec ''val'' comme valeur\n
+XMLInputSaxDialog.Column.Repeat.Label=R\u00E9p\u00E9ter
+XMLInputSaxDialog.Dialog.ErrorDisplayingPreviewData.Message=PDI a rencontr\u00E9 une rrreur durant la pr\u00E9visualisarion des donn\u00E9es
+XMLInputSaxDialog.Button.PreviewRows.Label=\  &Pr\u00E9visualiser lignes   
+XMLInputSaxDialog.Dialog.SelectVariable.Message=S\u00E9lectionner une variable d''environnement
+XMLInputSaxDialog.Button.EditEntry.ToolTip=Editer le fichier s\u00E9lectionn\u00E9 et le retirer de la liste.
+XMLInputSaxDialog.Button.Brows.ToolTip=Parcourir un r\u00E9pertoire et l''ajouter lui ou un fichier dans la liste
+XMLInputSaxDialog.IncludeFilename.Label=Inclure non de fichier dans le flux de sortie? 
+XMLInputSaxDialog.Column.DefiningAttribute.Label=D\u00E9finition attribut
+XMLInputSaxDialog.Button.AddFile.ToolTip=Ajouter cette entr\u00E9e \u00E0 la liste des fichiers et r\u00E9pertoires.
+XMLInputSaxDialog.Location.Label=Position (voir \u00E9galement\: aide infos bulle)
+XMLInputSaxDialog.FilenameFieldname.Label=Champs nom de fichier 
+XMLInputSaxDialog.Dialog.ErrorParsingInputData.Title=Erreur lors de l''analyse des donn\u00E9es
+XMLInputSaxDialog.System.Button.Browse=&Parcourir
+XMLInputSaxDialog.Button.DeleteEntry.ToolTip=Supprimer les entr\u00E9es s\u00E9lectionn\u00E9es depuis la liste.
+XMLInputSaxDialog.RowNumInOutputField.Label=Champ num\u00E9ro de ligne 
+XMLInputSaxDialog.Dialog.FilesRead.Message=Fichiers lus\:
+XMLInputSaxDialog.FileOrDirectory.Label=Fichier ou r\u00E9pertoire 
+XMLInputSaxDialog.Shell.Text=Flux XML en entr\u00E9e
+XMLInputSaxDialog.SelectedFiles.Label=Fichiers s\u00E9lectionn\u00E9s\: 
+XMLInputSaxDialog.Dialog.ErrorDisplayingPreviewData.Title=Erreur durant la pr\u00E9visualisarion des donn\u00E9es
+XMLInputSaxDialog.Column.Group.Label=Groupe
+XMLInputSaxDialog.File.Label=Fichier
+XMLInputSaxDialog.Column.Position.Label=Position
+XMLInputSaxDialog.Dialog.NoFilesFound.Title=ERREUR
+XMLInputSaxDialog.FileMaskRegExp.Label=Caract\u00E8res joker 
+XMLInputSaxDialog.Button.DeleteEntry.Label=&Supprimer
+XMLInputSaxDialog.Column.FileDirectory.ToolTip=Saisissez une extension ici et un nom de r\u00E9pertoire dans la premi\u00E8re colonne.
+XMLInputSaxDialog.Dialog.ErrorParsingInputData.Message=PDI a rencontr\u00E9 une erreur durant l''analyse des donn\u00E9es
+XMLInputSaxDialog.Tab.Fields.Label=Champs
+XMLInputSaxDialog.Button.ShowFilenames.Label=&Afficher nom de fichier(s)...
+XMLInputSaxDialog.Column.Type.Label=Type
+XMLInputSaxDialog.Dialog.NoFilesFound.Message=Aucun fichier trouv\u00E9\!  Veuillez svp v\u00E9rifier le nom de fichier/r\u00E9pertoire et les extensions.
+XMLInputSaxDialog.Limit.Label=Limite 
+XMLInputSaxDialog.Column.Currency.Label=Devise
+XMLInputSaxDialog.Column.Wildcard.Label=Extension (RegExp)
+XMLInputSaxDialog.Dialog.EnterPreviewSize.Message=Saisissez le nombre de lignes limite \u00E0 pr\u00E9visualiser\:
+XMLInputSaxDialog.68=Tous les fichiers
+XMLInputSaxDialog.System.Button.OK=\ &OK 
+XMLInputSaxDialog.RowNumInOutput.ToolTip=Option \u00E0 s\u00E9lectionner pour ajouter au flux de sortie, un champ contenant le nom du fichier trait\u00E9.
+XMLInputSaxDialog.Button.EditEntry.Label=&Editer
+XMLInputSaxDialog.Tab.Content.Label=Contenu
+XMLInputSaxDialog.Column.FileDirectory.Label=Fichier/R\u00E9pertoire
+XMLInputSaxDialog.RowNumInOutput.Label=Inclure num\u00E9ro de ligne dans le flux de sortie? 
+XMLInputSaxDialog.60=Fichier XML
+XMLInputSaxDialog.Dialog.FilesRead.Title=Fichiers lus
+XMLInputSaxDialog.Column.Element.Label=\u00C9l\u00E9ment
+XMLInputSaxDialog.Dialog.NoElements.Message=Aucun \u00E9l\u00E9ment d\u00E9fini.
+XMLInputSaxDialog.Dialog.EnterPreviewSize.Title=Saisissez une limite pour la pr\u00E9visualisation
+XMLInputSaxDialog.Column.Length.Label=Longeur
+XMLInputSaxDialog.IncludeFilename.ToolTip=Cochez cette option si vous souhaitez ajouter au flux de sortie, un champs (chaine de caract\u00E8re) contenant le nom du fichier trait\u00E9.
+XMLInputSaxDialog.Name=Extraction depuis flux XML (SAX)
+XMLInputSaxDialog.Description=Lire donn\u00E9es depuis un fichier XML avec SAX
