+++ plugins/xml-input-stream-plugin/core/src/main/resources/org/pentaho/di/trans/steps/xmlinputsax/messages/messages_it_IT.properties
@@ -1,0 +1,63 @@
+#File generated by Pentaho Translator for package 'org.pentaho.di.trans.steps.xmlinputsax' in locale 'it_IT'
+#
+#
+#Tue May 19 19:23:40 GMT 2009
+XMLInputSaxDialog.Dialog.NoElements.Title=Impossibile prelevare campi
+XMLInputSaxDialog.Column.Precision.Label=Precisione
+XMLInputSaxDialog.Button.AddFile.Label=&Aggiungi
+XMLInputSaxDialog.Column.Format.Label=Formato
+XMLInputSaxDialog.Button.GetFields.Label=&Preleva campi
+XMLInputSaxDialog.Column.Name.Label=Nome
+XMLInputSaxDialog.Column.Elements.Label=Elementi
+XMLInputSaxDialog.Column.Decimal.Label=Decimale
+XMLInputSaxDialog.StepName.Label=Nome del passo 
+XMLInputSaxDialog.Column.TrimType.Label=Tipo trim
+XMLInputSaxDialog.Location.ToolTip=Elemento/posizione\r\n------------------------\r\nA\: specificare un attributo\r\nEp\: specificare un elemento definito dalla sua posizione (equivalente ad E nell''originale XMLInput).\r\nEa\: specificare un elemento definito da un attributo e consentire analisi valore.\r\n\r\nEsempio\:\r\n-------------------\r\nEp\=elemento/1 questo \u00E8 il primo elemento chiamato ''elemento''\r\nEa\=elemento/att\:val questo \u00E8 l''elemento chiamato ''elemento'' che ha un attributo chiamato ''att'' con valore ''val''
+XMLInputSaxDialog.Column.Repeat.Label=Ripeti
+XMLInputSaxDialog.Dialog.ErrorDisplayingPreviewData.Message=Un errore si \u00E8 verificato durante l''anteprima dei dati XML
+XMLInputSaxDialog.Button.PreviewRows.Label=&Anteprima righe
+XMLInputSaxDialog.Button.EditEntry.ToolTip=Modifica il file selezionato e rimuovilo dalla lista.
+XMLInputSaxDialog.Button.Brows.ToolTip=Sfoglia file e cartelle e aggiungi alla lista
+XMLInputSaxDialog.IncludeFilename.Label=Includere il nome del file nell''output?
+XMLInputSaxDialog.Column.DefiningAttribute.Label=Definizione attributo
+XMLInputSaxDialog.Button.AddFile.ToolTip=Aggiungi quest''entry alla lista dei file e cartelle.
+XMLInputSaxDialog.Location.Label=Locazione (vedere anche\: aiuto suggerimento)
+XMLInputSaxDialog.FilenameFieldname.Label=Campo nome file
+XMLInputSaxDialog.Dialog.ErrorParsingInputData.Title=Errore durante l''analisi dei dati d''input
+XMLInputSaxDialog.System.Button.Browse=&Sfoglia
+XMLInputSaxDialog.Button.DeleteEntry.ToolTip=Cancella le entry selezionate dalla lista dei file.
+XMLInputSaxDialog.RowNumInOutputField.Label=campo con n\u00B0 riga
+XMLInputSaxDialog.Dialog.FilesRead.Message=File letti\: 
+XMLInputSaxDialog.FileOrDirectory.Label=File o cartella 
+XMLInputSaxDialog.Shell.Text=Streaming XML Input
+XMLInputSaxDialog.SelectedFiles.Label=File selezionati\: 
+XMLInputSaxDialog.Dialog.ErrorDisplayingPreviewData.Title=Errore durante la visualizzazione dell''anteprima dei dati
+XMLInputSaxDialog.File.Label=File
+XMLInputSaxDialog.Column.Group.Label=Gruppo
+XMLInputSaxDialog.Column.Position.Label=Posizione
+XMLInputSaxDialog.Dialog.NoFilesFound.Title=ERRORE
+XMLInputSaxDialog.FileMaskRegExp.Label=Espressione regolare 
+XMLInputSaxDialog.Button.DeleteEntry.Label=&Cancella
+XMLInputSaxDialog.Column.FileDirectory.ToolTip=Inserisci qua un''espressione regolare ed una cartella nella prima colonna.
+XMLInputSaxDialog.Dialog.ErrorParsingInputData.Message=Errore durante l''analisi dei dati d''input in questa finestra
+XMLInputSaxDialog.Tab.Fields.Label=Campi
+XMLInputSaxDialog.Button.ShowFilenames.Label=&Mostra file...
+XMLInputSaxDialog.Column.Type.Label=Tipo
+XMLInputSaxDialog.Dialog.NoFilesFound.Message=Nessun file trovato\! Prego controllare il file/cartella e le opzione dell''espressione regolare.
+XMLInputSaxDialog.Limit.Label=Limite 
+XMLInputSaxDialog.Column.Wildcard.Label=Wildcard (RegExp)
+XMLInputSaxDialog.Column.Currency.Label=Valuta
+XMLInputSaxDialog.Dialog.EnterPreviewSize.Message=Inserisci il numero di righe per l''anteprima\: 
+XMLInputSaxDialog.68=Tutti i file
+XMLInputSaxDialog.RowNumInOutput.ToolTip=Seleziona qua per aggiungere un campo (Stringa) contenente il nome del file.
+XMLInputSaxDialog.Button.EditEntry.Label=&Modifica
+XMLInputSaxDialog.Tab.Content.Label=Contenuto
+XMLInputSaxDialog.Column.FileDirectory.Label=File/Cartella
+XMLInputSaxDialog.RowNumInOutput.Label=n\u00B0 riga in output? 
+XMLInputSaxDialog.60=File XML
+XMLInputSaxDialog.Column.Element.Label=Elemento
+XMLInputSaxDialog.Dialog.FilesRead.Title=File letti
+XMLInputSaxDialog.Dialog.NoElements.Message=Nessun elemento definito.
+XMLInputSaxDialog.Dialog.EnterPreviewSize.Title=Inserisci la dimensione dell''anteprima
+XMLInputSaxDialog.Column.Length.Label=Lunghezza
+XMLInputSaxDialog.IncludeFilename.ToolTip=Seleziona qua per aggiungere un campo (Stringa) contenente il nome del file.
