+++ plugins/xml-input-stream-plugin/core/src/main/resources/org/pentaho/di/trans/steps/xmlinputsax/messages/messages_ko_KR.properties
@@ -1,0 +1,57 @@
+#File generated by Pentaho Translator for package 'org.pentaho.di.trans.steps.xmlinputsax' in locale 'ko_KR'
+#
+#
+#Tue Sep 01 13:56:27 KST 2009
+XMLInputSaxDialog.Dialog.NoElements.Title=\uD544\uB4DC\uB97C \uAC00\uC838\uC62C \uC218 \uC5C6\uC2B5\uB2C8\uB2E4
+XMLInputSaxDialog.Column.Precision.Label=\uC815\uBC00\uB3C4
+XMLInputSaxDialog.Button.AddFile.Label=\uCD94\uAC00(&A)
+XMLInputSaxDialog.Column.Format.Label=\uD615\uC2DD
+XMLInputSaxDialog.Button.GetFields.Label=\uD544\uB4DC \uAC00\uC838\uC624\uAE30(&G)
+XMLInputSaxDialog.Column.Name.Label=\uC774\uB984
+XMLInputSaxDialog.Column.Decimal.Label=Decimal
+XMLInputSaxDialog.StepName.Label=Step \uC774\uB984 
+XMLInputSaxDialog.Column.TrimType.Label=Trim \uD615\uC2DD
+XMLInputSaxDialog.Column.Repeat.Label=\uBC18\uBCF5
+XMLInputSaxDialog.Dialog.ErrorDisplayingPreviewData.Message=XML \uB370\uC774\uD130\uB97C \uBBF8\uB9AC\uBCF4\uAE30 \uD558\uB294 \uB3D9\uC548 \uC624\uB958\uAC00 \uBC1C\uC0DD\uD558\uC600\uC2B5\uB2C8\uB2E4
+XMLInputSaxDialog.Button.PreviewRows.Label=\ \uB85C\uC6B0 \uBBF8\uB9AC\uBCF4\uAE30(&P)
+XMLInputSaxDialog.Button.EditEntry.ToolTip=\uC120\uD0DD\uD55C \uD30C\uC77C\uC744 \uD3B8\uC9D1\uD558\uACE0 \uB9AC\uC2A4\uD2B8\uC5D0\uC11C \uC81C\uAC70
+XMLInputSaxDialog.Button.Brows.ToolTip=\uB9AC\uC2A4\uD2B8\uC5D0 \uCD94\uAC00\uD560 \uD30C\uC77C \uB610\uB294 \uB514\uB809\uD1A0\uB9AC \uCC3E\uC544\uBCF4\uAE30
+XMLInputSaxDialog.IncludeFilename.Label=\uCD9C\uB825\uC5D0 \uD30C\uC77C\uC774\uB984 \uD3EC\uD568?
+XMLInputSaxDialog.Column.DefiningAttribute.Label=Defining Attribute
+XMLInputSaxDialog.Button.AddFile.ToolTip=\uC774 \uD56D\uBAA9\uC744 \uD30C\uC77C\uACFC \uB514\uB809\uD1A0\uB9AC \uB9AC\uC2A4\uD2B8\uC5D0 \uCD94\uAC00
+XMLInputSaxDialog.Location.Label=\uC704\uCE58 (\uD234\uD301 \uB3C4\uC6B0\uB9D0 \uCC38\uC870)
+XMLInputSaxDialog.FilenameFieldname.Label=\uD30C\uC77C\uC774\uB984 \uD544\uB4DC\uC774\uB984
+XMLInputSaxDialog.Dialog.ErrorParsingInputData.Title=\uC785\uB825 \uB370\uC774\uD130 \uD30C\uC2F1 \uC624\uB958
+XMLInputSaxDialog.System.Button.Browse=\uCC3E\uC544\uBCF4\uAE30(&B)
+XMLInputSaxDialog.Button.DeleteEntry.ToolTip=\uD30C\uC77C \uB9AC\uC2A4\uD2B8\uC758 \uC120\uD0DD\uB41C \uD56D\uBAA9\uC744 \uC0AD\uC81C.
+XMLInputSaxDialog.RowNumInOutputField.Label=\uB85C\uC6B0\uBC88\uD638 \uD544\uB4DC\uC774\uB984
+XMLInputSaxDialog.Dialog.FilesRead.Message=\uC77D\uC740 \uD30C\uC77C\:
+XMLInputSaxDialog.FileOrDirectory.Label=\uD30C\uC77C \uB610\uB294 \uB514\uB809\uD1A0\uB9AC
+XMLInputSaxDialog.Shell.Text=Streaming XML Input
+XMLInputSaxDialog.SelectedFiles.Label=\uC120\uD0DD\uD55C \uD30C\uC77C\:
+XMLInputSaxDialog.Dialog.ErrorDisplayingPreviewData.Title=\uBBF8\uB85C\uBCF4\uAE30 \uB370\uC774\uD130 \uD45C\uC2DC \uC624\uB958
+XMLInputSaxDialog.File.Label=\uD30C\uC77C
+XMLInputSaxDialog.Column.Group.Label=\uADF8\uB8F9
+XMLInputSaxDialog.Column.Position.Label=\uC704\uCE58
+XMLInputSaxDialog.Dialog.NoFilesFound.Title=\uC624\uB958
+XMLInputSaxDialog.FileMaskRegExp.Label=\uC815\uADDC \uD45C\uD604\uC2DD
+XMLInputSaxDialog.Button.DeleteEntry.Label=\uC0AD\uC81C(&D)
+XMLInputSaxDialog.Tab.Fields.Label=\uD544\uB4DC
+XMLInputSaxDialog.Button.ShowFilenames.Label=\uD30C\uC77C\uC774\uB984 \uBCF4\uAE30(&S)...
+XMLInputSaxDialog.Column.Type.Label=\uB370\uC774\uD130\uD615
+XMLInputSaxDialog.Limit.Label=Limit 
+XMLInputSaxDialog.Dialog.NoFilesFound.Message=\uD30C\uC77C\uC744 \uCC3E\uC744 \uC218 \uC5C6\uC2B5\uB2C8\uB2E4\! \uD30C\uC77C\uC774\uB984/\uB514\uB809\uD1A0\uB9AC \uADF8\uB9AC\uACE0 \uC815\uADDC\uD45C\uD604\uC2DD \uC635\uC158\uC744 \uD655\uC778\uD558\uC2ED\uC2DC\uC624.
+XMLInputSaxDialog.Column.Currency.Label=\uD1B5\uD654
+XMLInputSaxDialog.Column.Wildcard.Label=Wildcard
+XMLInputSaxDialog.Dialog.EnterPreviewSize.Message=\uBBF8\uB9AC\uBCF4\uAE30\uD560 \uB85C\uC6B0\uC758 \uAC1C\uC218\uB97C \uC785\uB825\:
+XMLInputSaxDialog.68=\uBAA8\uB4E0 \uD30C\uC77C
+XMLInputSaxDialog.Button.EditEntry.Label=\uD3B8\uC9D1(&E)
+XMLInputSaxDialog.Tab.Content.Label=\uB0B4\uC6A9
+XMLInputSaxDialog.Column.FileDirectory.Label=\uD30C\uC77C/\uB514\uB809\uD1A0\uB9AC
+XMLInputSaxDialog.RowNumInOutput.Label=\uCD9C\uB825\uC5D0 \uB85C\uC6B0\uBC88\uD638?
+XMLInputSaxDialog.Column.Element.Label=Element
+XMLInputSaxDialog.60=XML \uD30C\uC77C
+XMLInputSaxDialog.Dialog.FilesRead.Title=\uC77D\uC740 \uD30C\uC77C
+XMLInputSaxDialog.Dialog.NoElements.Message=\uC5D8\uB9AC\uBA3C\uD2B8\uAC00 \uC815\uC758\uB418\uC9C0 \uC54A\uC558\uC2B5\uB2C8\uB2E4.
+XMLInputSaxDialog.Dialog.EnterPreviewSize.Title=\uBBF8\uB9AC\uBCF4\uAE30 \uAC1C\uC218 \uC785\uB825
+XMLInputSaxDialog.Column.Length.Label=\uAE38\uC774
