+++ plugins/xml-input-stream-plugin/core/src/main/resources/version.properties
@@ -1,0 +1,4 @@
+${build.description} build information
+
+version=${build.revision}
+builddate=${timestamp}
