--- plugins/xml-input-stream/build.properties
@@ -1,10 +1,0 @@
-project.revision=8.0-SNAPSHOT
-ivy.artifact.group=pentaho-kettle
-ivy.artifact.id=xml-input-stream-plugin
-impl.vendor=Pentaho Corporation
-impl.title=XML Input Stream Plugin
-impl.productID=${ivy.artifact.id}
-dev-lib.dir=dev-lib
-
-dependency.kettle.revision=8.0-SNAPSHOT
-dependency.pentaho-xul.revision=8.0-SNAPSHOT
