--- plugins/xml-input-stream/ivysettings.xml
@@ -1,31 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<ivysettings>
-  <properties environment="env" />
-  <property name="maven.local.repo" value="${user.home}/.m2/repository" override="false" />
-  <property name="ivy.local.default.root" value="${ivy.default.ivy.user.dir}/local" override="true" />
-  <property name="ivy.local.default.artifact.pattern" value="[organisation]/[module]/[revision]/[type]s/[artifact]-[revision].[ext]"
-    override="false" />
-    
-  <!-- Repository for Pentaho-hosted artifacts -->  
-  <property name="pentaho.resolve.repo" value="http://ivy-nexus.pentaho.org/content/groups/omni" override="false" />
-  <!-- Repository for External-hosted artifacts (Optional. Defaults to Pentaho-hosted.) -->
-  <property name="public.resolve.repo" value="${pentaho.resolve.repo}" override="false" />
-
-  <settings defaultResolver="pentaho-chained-resolver" />
-  <include url="${ivy.default.settings.dir}/ivysettings-public.xml" />
-  <include url="${ivy.default.settings.dir}/ivysettings-local.xml" />
-  <resolvers>
-    <chain name="pentaho-chained-resolver">
-      <resolver ref="local" />
-      <dual name="pentaho">
-        <url name="pentaho-ivy" checkmodified="true" changingPattern=".*-SNAPSHOT.*">
-          <ivy pattern="${pentaho.resolve.repo}/[organisation]/[module]/[revision]/[module]-[revision].ivy.xml" />
-        </url>
-        <ibiblio name="pentaho-mvn" m2compatible="true" root="${pentaho.resolve.repo}" checkmodified="true" changingPattern=".*-SNAPSHOT.*" />
-      </dual>
-      <ibiblio name="local-mvn" m2compatible="true" root="file:///${maven.local.repo}" checkmodified="true" changingPattern=".*-SNAPSHOT.*" />
-      <ibiblio name="public-maven" root="${public.resolve.repo}" m2compatible="true" />
-    </chain>
-  </resolvers>
-  <caches lockStrategy="artifact-lock" resolutionCacheDir="${ivy.default.ivy.user.dir}/resol-cache${env.EXECUTOR_NUMBER}" />
-</ivysettings>
