--- plugins/xml-input-stream/package-ivy.xml
@@ -1,17 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<ivy-module version="2.0" xmlns:m="http://ant.apache.org/ivy/maven">
-  <info organisation="${ivy.artifact.group}" module="${ivy.artifact.id}" revision="${project.revision}" />
-  <configurations>
-    <conf name="default" />
-  </configurations>
-
-  <publications>
-    <artifact name="${ivy.artifact.id}" type="zip" />
-    <artifact name="${ivy.artifact.id}" type="jar" />
-    <artifact name="${ivy.artifact.id}" m:classifier="sources" type="source" ext="jar"/>
-    <artifact name="${ivy.artifact.id}" m:classifier="sources" type="source" ext="zip"/>
-  </publications>
-
-  <dependencies defaultconf="default->default">
-  </dependencies>
-</ivy-module>
