--- plugins/xml-input-stream/res/version.xml
@@ -1,2 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<version branch='TRUNK'>@TRUNK@</version>
